4.1, 4.2

+++ lesson4/4.1.py
@@ -3,7 +3,7 @@ from sys import argv
 
 def salary():
     try:
-        time, rate, bonus = map(float, argv[1:])
+        time, rate, bonus = map(int, argv[1:])
         print(time * rate + bonus)
     except ValueError:
         print('Вводите три числа!')
